Sync pes (#3)
* Synchronous PES mode

The PES mode encapsulation supports now the Synchronous mode.

This mode includes PTS marks in the PES packets. However these marks are only useful for external tools. So, don't use this mode except if you need to remux with an external tool that **requires** to have PTS in Metadata packets. In any other case, it's a waste of space.

* Example: Reverse order of add-programinfo-id

* Use correct values